requirement.txt
asgiref==3.3.4
autopep8==1.5.7
dateutils==0.6.12
Django==3.2.4
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
djangorestframework==3.12.4
flake8==3.9.2
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
